refactor: Renaming the top-level modules
