+++ NOTE.md
@@ -3,6 +3,7 @@
 - [ ] Load diagram from file
 - [ ] Save diagram as xml
 - [ ] Save diagram as svg
+- [ ] Import color picker (see experiments project) + css
 - [ ] Basic language constructs
   - [ ] Evidence Source
   - [ ] Evidence Type
